+++ frontend/src/UI/components/Buttons/EditButton.tsx
@@ -1,14 +1,65 @@
 import { EditIcon } from "@chakra-ui/icons";
 import {
+  Button,
+  ButtonGroup,
+  FormControl,
+  FormLabel,
   IconButton,
+  Input,
   Popover,
   PopoverArrow,
   PopoverCloseButton,
   PopoverContent,
   PopoverTrigger,
+  Stack,
   useDisclosure,
 } from "@chakra-ui/react";
-import { EditForm } from "../Form/EditForm";
+import { useState } from "react";
+
+function EditForm({
+  id,
+  onCancel,
+  onClose,
+  formLabel,
+  initialValue,
+  updateValue,
+}: {
+  id: string;
+  onCancel: () => void;
+  onClose: () => void;
+  formLabel: string;
+  initialValue: string;
+  updateValue: (payload: {
+    id: string;
+    name: string;
+  }) => Promise<Promise<void>>;
+}) {
+  const [value, setValue] = useState(initialValue);
+
+  const submit = () => {
+    updateValue({ id, name: value });
+    onClose();
+  };
+
+  return (
+    <Stack spacing={4}>
+      <FormControl>
+        <FormLabel htmlFor="inputField">{formLabel}</FormLabel>
+        <Input
+          id="inputField"
+          value={value}
+          onChange={(e) => setValue(e.target.value)}
+        />
+      </FormControl>
+      <ButtonGroup display="flex" justifyContent="flex-end">
+        <Button variant="outline" onClick={onCancel}>
+          Cancel
+        </Button>
+        <Button onClick={submit}>Save</Button>
+      </ButtonGroup>
+    </Stack>
+  );
+}
 
 type Props = {
   id: string;
